requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
pycodestyle==2.11.0
pydocstyle==6.3.0
python-dateutil==2.8.2
python-docx==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
